+++ src/main/java/com/kyrakova/diploma/models/model/Model.java
@@ -13,7 +13,7 @@ public class Model extends BaseModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    public int getId() {
+    public Long getId() {
         return id;
     }
